+++ tests/queries/test_event_query.py
@@ -388,3 +388,16 @@ def test_subquery_handles_nested_subquery():
     assert isinstance(q.groups[0].subgroups[0].subgroups[0], FilterGroup)
     assert q.groups[0].subgroups[0].subgroups[0].filters[0].term == "term"
     assert q.groups[0].subgroups[0].subgroups[0].filters[0].value == "value"
+
+def test_on_filter_creates_correct_filter():
+    q = EventQuery().on(term="date_term", date=datetime(2025,1,1,1,1,1,1))
+    expected = FilterGroup(
+        filters=[
+            Filter(
+                term="date_term",
+                operator="ON",
+                value="2025-01-01",
+            )
+        ]
+    )
+    assert q.groups.pop() == expected